requirements.txt
Twisted==14.0.2
argparse==1.2.1
enum==0.4.4
smpp.pdu==0.3
wsgiref==0.1.2
zope.interface==4.1.1
chardet